// web/src/shared/api/generated/model/lessonDetailResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LessonObjectiveResult } from './lessonObjectiveResult';

export interface LessonDetailResult {
  id: string;
  unitId: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  order: number | string;
  state: string;
  explanation: string;
  summary: string;
  /** @nullable */
  videoUrl: string | null;
  objectives: LessonObjectiveResult[];
}